requirements.txt
python-telegram-bot==20.7
openai==1.30.1
redis==5.0.1
mysql-connector-python==8.4.0
requests==2.31.0
